move DOVIInfo to new avdesc package
